+++ Source/PINDiskCache.m
@@ -9,6 +9,7 @@
 #endif
 
 #import <pthread.h>
+#import <sys/xattr.h>
 
 #import <PINOperation/PINOperation.h>
 
@@ -18,6 +19,10 @@ __LINE__, [error localizedDescription]); }
 
 #define PINDiskCacheException(exception) if (exception) { NSAssert(NO, [exception reason]); }
 
+const char * PINDiskCacheAgeLimitAttributeName = "com.pinterest.PINDiskCache.ageLimit";
+NSString * const PINDiskCacheErrorDomain = @"com.pinterest.PINDiskCache";
+NSErrorUserInfoKey const PINDiskCacheErrorReadFailureCodeKey = @"PINDiskCacheErrorReadFailureCodeKey";
+NSErrorUserInfoKey const PINDiskCacheErrorWriteFailureCodeKey = @"PINDiskCacheErrorWriteFailureCodeKey";
 NSString * const PINDiskCachePrefix = @"com.pinterest.PINDiskCache";
 static NSString * const PINDiskCacheSharedName = @"PINDiskCacheShared";
 
@@ -40,9 +45,22 @@ static PINOperationDataCoalescingBlock PINDiskTrimmingDateCoalescingBlock = ^id(
     return (result == NSOrderedDescending) ? newDate : existingDate;
 };
 
+const char * PINDiskCacheFileSystemRepresentation(NSURL *url)
+{
+#ifdef __MAC_10_13 // Xcode >= 9
+    // -fileSystemRepresentation is available on macOS >= 10.9
+    if (@available(macOS 10.9, iOS 7.0, watchOS 2.0, tvOS 9.0, *)) {
+      return url.fileSystemRepresentation;
+    }
+#endif
+    return [url.path cStringUsingEncoding:NSUTF8StringEncoding];
+}
+
 @interface PINDiskCacheMetadata : NSObject
-@property (nonatomic, strong) NSDate *date;
+@property (nonatomic, strong) NSDate *createdDate;
+@property (nonatomic, strong) NSDate *lastModifiedDate;
 @property (nonatomic, strong) NSNumber *size;
+@property (nonatomic) NSTimeInterval ageLimit;
 @end
 
 @interface PINDiskCache () {
@@ -144,6 +162,26 @@ static NSURL *_sharedTrashURL;
                   keyDecoder:(PINDiskCacheKeyDecoderBlock)keyDecoder
               operationQueue:(PINOperationQueue *)operationQueue
 {
+    return [self initWithName:name prefix:prefix
+                     rootPath:rootPath
+                   serializer:serializer
+                 deserializer:deserializer
+                   keyEncoder:keyEncoder
+                   keyDecoder:keyDecoder
+               operationQueue:operationQueue
+                     ttlCache:NO];
+}
+
+- (instancetype)initWithName:(NSString *)name
+                      prefix:(NSString *)prefix
+                    rootPath:(NSString *)rootPath
+                  serializer:(PINDiskCacheSerializerBlock)serializer
+                deserializer:(PINDiskCacheDeserializerBlock)deserializer
+                  keyEncoder:(PINDiskCacheKeyEncoderBlock)keyEncoder
+                  keyDecoder:(PINDiskCacheKeyDecoderBlock)keyDecoder
+              operationQueue:(PINOperationQueue *)operationQueue
+                    ttlCache:(BOOL)ttlCache
+{
     if (!name)
         return nil;
     
@@ -161,6 +199,7 @@ static NSURL *_sharedTrashURL;
         _name = [name copy];
         _prefix = [prefix copy];
         _operationQueue = operationQueue;
+        _ttlCache = ttlCache;
         _willAddObjectBlock = nil;
         _willRemoveObjectBlock = nil;
         _willRemoveAllObjectsBlock = nil;
@@ -453,7 +492,7 @@ static NSURL *_sharedTrashURL;
 - (void)initializeDiskProperties
 {
     NSUInteger byteCount = 0;
-    NSArray *keys = @[ NSURLContentModificationDateKey, NSURLTotalFileAllocatedSizeKey ];
+    NSArray *keys = @[ NSURLCreationDateKey, NSURLContentModificationDateKey, NSURLTotalFileAllocatedSizeKey ];
     
     NSError *error = nil;
     
@@ -479,15 +518,34 @@ static NSURL *_sharedTrashURL;
             if (_metadata[key] == nil) {
                 _metadata[key] = [[PINDiskCacheMetadata alloc] init];
             }
+
+            NSDate *createdDate = [dictionary objectForKey:NSURLCreationDateKey];
+            if (createdDate && key)
+                _metadata[key].createdDate = createdDate;
         
-            NSDate *date = [dictionary objectForKey:NSURLContentModificationDateKey];
-            if (date && key)
-                _metadata[key].date = date;
+            NSDate *lastModifiedDate = [dictionary objectForKey:NSURLContentModificationDateKey];
+            if (lastModifiedDate && key)
+                _metadata[key].lastModifiedDate = lastModifiedDate;
         
             NSNumber *fileSize = [dictionary objectForKey:NSURLTotalFileAllocatedSizeKey];
             if (fileSize) {
                 _metadata[key].size = fileSize;
                 byteCount += [fileSize unsignedIntegerValue];
+            }
+
+            if (_ttlCache) {
+                NSTimeInterval ageLimit;
+                ssize_t res = getxattr(PINDiskCacheFileSystemRepresentation(fileURL), PINDiskCacheAgeLimitAttributeName, &ageLimit, sizeof(NSTimeInterval), 0, 0);
+                if(res) {
+                    _metadata[key].ageLimit = ageLimit;
+                } else if (res == -1) {
+                    // Ignore if the extended attribute was never recorded for this file.
+                    if (errno != ENOATTR) {
+                      NSDictionary<NSErrorUserInfoKey, id> *userInfo = @{ PINDiskCacheErrorReadFailureCodeKey : @(errno)};
+                      error = [NSError errorWithDomain:PINDiskCacheErrorDomain code:PINDiskCacheErrorReadFailure userInfo:userInfo];
+                      PINDiskCacheError(error);
+                    }
+                }
             }
         [self unlock];
     }
@@ -498,6 +556,9 @@ static NSURL *_sharedTrashURL;
     
         if (self->_byteLimit > 0 && self->_byteCount > self->_byteLimit)
             [self trimToSizeByDateAsync:self->_byteLimit completion:nil];
+
+        if (self->_ttlCache)
+            [self removeExpiredObjectsAsync:nil];
     
         _diskStateKnown = YES;
         pthread_cond_broadcast(&_diskStateKnownCondition);
@@ -528,11 +589,54 @@ static NSURL *_sharedTrashURL;
     if (success) {
         NSString *key = [self keyForEncodedFileURL:fileURL];
         if (key) {
-            _metadata[key].date = date;
+            _metadata[key].lastModifiedDate = date;
         }
     }
     
     return success;
+}
+
+- (void)asynchronouslySetAgeLimit:(NSTimeInterval)ageLimit forURL:(NSURL *)fileURL
+{
+    [self.operationQueue scheduleOperation:^{
+        [self lockForWriting];
+            [self _locked_setAgeLimit:ageLimit forURL:fileURL];
+        [self unlock];
+    } withPriority:PINOperationQueuePriorityLow];
+}
+
+- (BOOL)_locked_setAgeLimit:(NSTimeInterval)ageLimit forURL:(NSURL *)fileURL
+{
+    if (!fileURL) {
+        return NO;
+    }
+
+    NSError *error = nil;
+    if (ageLimit <= 0.0) {
+        if (removexattr(PINDiskCacheFileSystemRepresentation(fileURL), PINDiskCacheAgeLimitAttributeName, 0) != 0) {
+          // Ignore if the extended attribute was never recorded for this file.
+          if (errno != ENOATTR) {
+            NSDictionary<NSErrorUserInfoKey, id> *userInfo = @{ PINDiskCacheErrorWriteFailureCodeKey : @(errno)};
+            error = [NSError errorWithDomain:PINDiskCacheErrorDomain code:PINDiskCacheErrorWriteFailure userInfo:userInfo];
+            PINDiskCacheError(error);
+          }
+        }
+    } else {
+        if (setxattr(PINDiskCacheFileSystemRepresentation(fileURL), PINDiskCacheAgeLimitAttributeName, &ageLimit, sizeof(NSTimeInterval), 0, 0) != 0) {
+            NSDictionary<NSErrorUserInfoKey, id> *userInfo = @{ PINDiskCacheErrorWriteFailureCodeKey : @(errno)};
+            error = [NSError errorWithDomain:PINDiskCacheErrorDomain code:PINDiskCacheErrorWriteFailure userInfo:userInfo];
+            PINDiskCacheError(error);
+        }
+    }
+
+    if (!error) {
+        NSString *key = [self keyForEncodedFileURL:fileURL];
+        if (key) {
+            _metadata[key].ageLimit = ageLimit;
+        }
+    }
+
+    return !error;
 }
 
 - (BOOL)removeFileAndExecuteBlocksForKey:(NSString *)key
@@ -612,18 +716,22 @@ static NSURL *_sharedTrashURL;
 
 - (void)trimDiskToSizeByDate:(NSUInteger)trimByteCount
 {
+    if (self.isTTLCache) {
+        [self removeExpiredObjects];
+    }
+
     NSMutableArray *keysToRemove = nil;
-    
+  
     [self lockForWriting];
         if (_byteCount > trimByteCount) {
             keysToRemove = [[NSMutableArray alloc] init];
             
-            NSArray *keysSortedByDate = [_metadata keysSortedByValueUsingComparator:^NSComparisonResult(PINDiskCacheMetadata * _Nonnull obj1, PINDiskCacheMetadata * _Nonnull obj2) {
-                return [obj1.date compare:obj2.date];
+            NSArray *keysSortedByLastModifiedDate = [_metadata keysSortedByValueUsingComparator:^NSComparisonResult(PINDiskCacheMetadata * _Nonnull obj1, PINDiskCacheMetadata * _Nonnull obj2) {
+                return [obj1.lastModifiedDate compare:obj2.lastModifiedDate];
             }];
             
             NSUInteger bytesSaved = 0;
-            for (NSString *key in keysSortedByDate) { // oldest objects first
+            for (NSString *key in keysSortedByLastModifiedDate) { // oldest objects first
                 [keysToRemove addObject:key];
                 NSNumber *byteSize = _metadata[key].size;
                 if (byteSize) {
@@ -644,18 +752,18 @@ static NSURL *_sharedTrashURL;
 - (void)trimDiskToDate:(NSDate *)trimDate
 {
     [self lockForWriting];
-        NSArray *keysSortedByDate = [_metadata keysSortedByValueUsingComparator:^NSComparisonResult(PINDiskCacheMetadata * _Nonnull obj1, PINDiskCacheMetadata * _Nonnull obj2) {
-            return [obj1.date compare:obj2.date];
+        NSArray *keysSortedByCreatedDate = [_metadata keysSortedByValueUsingComparator:^NSComparisonResult(PINDiskCacheMetadata * _Nonnull obj1, PINDiskCacheMetadata * _Nonnull obj2) {
+            return [obj1.createdDate compare:obj2.createdDate];
         }];
     
         NSMutableArray *keysToRemove = [[NSMutableArray alloc] init];
         
-        for (NSString *key in keysSortedByDate) { // oldest files first
-            NSDate *accessDate = _metadata[key].date;
-            if (!accessDate)
+        for (NSString *key in keysSortedByCreatedDate) { // oldest files first
+            NSDate *createdDate = _metadata[key].createdDate;
+            if (!createdDate || _metadata[key].ageLimit > 0.0)
                 continue;
             
-            if ([accessDate compare:trimDate] == NSOrderedAscending) { // older than trim date
+            if ([createdDate compare:trimDate] == NSOrderedAscending) { // older than trim date
                 [keysToRemove addObject:key];
             } else {
                 break;
@@ -739,9 +847,14 @@ static NSURL *_sharedTrashURL;
 
 - (void)setObjectAsync:(id <NSCoding>)object forKey:(NSString *)key completion:(PINDiskCacheObjectBlock)block
 {
+    [self setObjectAsync:object forKey:key withAgeLimit:0.0 completion:(PINDiskCacheObjectBlock)block];
+}
+
+- (void)setObjectAsync:(id <NSCoding>)object forKey:(NSString *)key withAgeLimit:(NSTimeInterval)ageLimit completion:(nullable PINDiskCacheObjectBlock)block
+{
     [self.operationQueue scheduleOperation:^{
         NSURL *fileURL = nil;
-        [self setObject:object forKey:key fileURL:&fileURL];
+        [self setObject:object forKey:key withAgeLimit:ageLimit fileURL:&fileURL];
         
         if (block) {
             block(self, key, object);
@@ -752,6 +865,11 @@ static NSURL *_sharedTrashURL;
 - (void)setObjectAsync:(id <NSCoding>)object forKey:(NSString *)key withCost:(NSUInteger)cost completion:(nullable PINCacheObjectBlock)block
 {
     [self setObjectAsync:object forKey:key completion:(PINDiskCacheObjectBlock)block];
+}
+
+- (void)setObjectAsync:(id <NSCoding>)object forKey:(NSString *)key withCost:(NSUInteger)cost ageLimit:(NSTimeInterval)ageLimit completion:(nullable PINCacheObjectBlock)block
+{
+    [self setObjectAsync:object forKey:key withAgeLimit:ageLimit completion:(PINDiskCacheObjectBlock)block];
 }
 
 - (void)removeObjectForKeyAsync:(NSString *)key completion:(PINDiskCacheObjectBlock)block
@@ -829,6 +947,17 @@ static NSURL *_sharedTrashURL;
                                 completion:completion];
 }
 
+- (void)removeExpiredObjectsAsync:(PINCacheBlock)block
+{
+    [self.operationQueue scheduleOperation:^{
+        [self removeExpiredObjects];
+
+        if (block) {
+            block(self);
+        }
+    } withPriority:PINOperationQueuePriorityLow];
+}
+
 - (void)removeAllObjectsAsync:(PINCacheBlock)block
 {
     [self.operationQueue scheduleOperation:^{
@@ -866,8 +995,13 @@ static NSURL *_sharedTrashURL;
 {
     [self lock];
         if (_metadata[key] != nil || _diskStateKnown == NO) {
+            BOOL objectExpired = NO;
+            if (self->_ttlCache && _metadata[key].createdDate != nil) {
+                NSTimeInterval ageLimit = _metadata[key].ageLimit > 0.0 ? _metadata[key].ageLimit : self->_ageLimit;
+                objectExpired = ageLimit > 0 && fabs([_metadata[key].createdDate timeIntervalSinceDate:[NSDate date]]) > ageLimit;
+            }
             [self unlock];
-            return ([self fileURLForKey:key updateFileModificationDate:NO] != nil);
+            return (!objectExpired && [self fileURLForKey:key updateFileModificationDate:NO] != nil);
         }
     [self unlock];
     return NO;
@@ -895,15 +1029,16 @@ static NSURL *_sharedTrashURL;
     id <NSCoding> object = nil;
     NSURL *fileURL = [self encodedFileURLForKey:key];
     
-    NSDate *now = [[NSDate alloc] init];
+    NSDate *now = [NSDate date];
     [self lock];
         if (self->_ttlCache) {
             // We actually need to know the entire disk state if we're a TTL cache.
             [self unlock];
             [self lockAndWaitForKnownState];
         }
-    
-        if (!self->_ttlCache || self->_ageLimit <= 0 || fabs([_metadata[key].date timeIntervalSinceDate:now]) < self->_ageLimit) {
+
+        NSTimeInterval ageLimit = _metadata[key].ageLimit > 0.0 ? _metadata[key].ageLimit : self->_ageLimit;
+        if (!self->_ttlCache || ageLimit <= 0 || fabs([_metadata[key].createdDate timeIntervalSinceDate:now]) < ageLimit) {
             // If the cache should behave like a TTL cache, then only fetch the object if there's a valid ageLimit and  the object is still alive
             
             NSData *objectData = [[NSData alloc] initWithContentsOfFile:[fileURL path]];
@@ -924,9 +1059,8 @@ static NSURL *_sharedTrashURL;
               }
               [self lock];
             }
-            if (object && !self->_ttlCache) {
+            if (object)
                 [self asynchronouslySetFileModificationDate:now forURL:fileURL];
-            }
         }
     [self unlock];
     
@@ -950,7 +1084,7 @@ static NSURL *_sharedTrashURL;
         return nil;
     }
     
-    NSDate *now = [[NSDate alloc] init];
+    NSDate *now = [NSDate date];
     NSURL *fileURL = [self encodedFileURLForKey:key];
     
     [self lockForWriting];
@@ -967,7 +1101,17 @@ static NSURL *_sharedTrashURL;
 
 - (void)setObject:(id <NSCoding>)object forKey:(NSString *)key
 {
-    [self setObject:object forKey:key fileURL:nil];
+    [self setObject:object forKey:key withAgeLimit:0.0];
+}
+
+- (void)setObject:(id <NSCoding>)object forKey:(NSString *)key withAgeLimit:(NSTimeInterval)ageLimit
+{
+    [self setObject:object forKey:key withAgeLimit:ageLimit fileURL:nil];
+}
+
+- (void)setObject:(id <NSCoding>)object forKey:(NSString *)key withCost:(NSUInteger)cost ageLimit:(NSTimeInterval)ageLimit
+{
+    [self setObject:object forKey:key withAgeLimit:ageLimit];
 }
 
 - (void)setObject:(id <NSCoding>)object forKey:(NSString *)key withCost:(NSUInteger)cost
@@ -984,8 +1128,10 @@ static NSURL *_sharedTrashURL;
     }
 }
 
-- (void)setObject:(id <NSCoding>)object forKey:(NSString *)key fileURL:(NSURL **)outFileURL
+- (void)setObject:(id <NSCoding>)object forKey:(NSString *)key withAgeLimit:(NSTimeInterval)ageLimit fileURL:(NSURL **)outFileURL
 {
+    NSAssert(ageLimit <= 0.0 || (ageLimit > 0.0 && _ttlCache), @"ttlCache must be set to YES if setting an object-level age limit.");
+
     if (!key || !object)
         return;
     
@@ -1031,7 +1177,7 @@ static NSURL *_sharedTrashURL;
             }
             
             NSError *error = nil;
-            NSDictionary *values = [fileURL resourceValuesForKeys:@[ NSURLContentModificationDateKey, NSURLTotalFileAllocatedSizeKey ] error:&error];
+            NSDictionary *values = [fileURL resourceValuesForKeys:@[ NSURLCreationDateKey, NSURLContentModificationDateKey, NSURLTotalFileAllocatedSizeKey ] error:&error];
             PINDiskCacheError(error);
             
             NSNumber *diskFileSize = [values objectForKey:NSURLTotalFileAllocatedSizeKey];
@@ -1043,11 +1189,15 @@ static NSURL *_sharedTrashURL;
                 self->_metadata[key].size = diskFileSize;
                 self.byteCount = self->_byteCount + [diskFileSize unsignedIntegerValue]; // atomic
             }
-            NSDate *date = [values objectForKey:NSURLContentModificationDateKey];
-            if (date) {
-                self->_metadata[key].date = date;
+            NSDate *createdDate = [values objectForKey:NSURLCreationDateKey];
+            if (createdDate) {
+                self->_metadata[key].createdDate = createdDate;
             }
-            
+            NSDate *lastModifiedDate = [values objectForKey:NSURLContentModificationDateKey];
+            if (lastModifiedDate) {
+                self->_metadata[key].lastModifiedDate = lastModifiedDate;
+            }
+            [self asynchronouslySetAgeLimit:ageLimit forURL:fileURL];
             if (self->_byteLimit > 0 && self->_byteCount > self->_byteLimit)
                 [self trimToSizeByDateAsync:self->_byteLimit completion:nil];
         } else {
@@ -1121,6 +1271,26 @@ static NSURL *_sharedTrashURL;
     [self trimDiskToSizeByDate:trimByteCount];
 }
 
+- (void)removeExpiredObjects
+{
+    [self lockForWriting];
+        NSDate *now = [NSDate date];
+        NSMutableArray<NSString *> *expiredObjectKeys = [NSMutableArray array];
+        [_metadata enumerateKeysAndObjectsUsingBlock:^(NSString * _Nonnull key, PINDiskCacheMetadata * _Nonnull obj, BOOL * _Nonnull stop) {
+            NSTimeInterval ageLimit = obj.ageLimit > 0.0 ? obj.ageLimit : self->_ageLimit;
+            NSDate *expirationDate = [obj.createdDate dateByAddingTimeInterval:ageLimit];
+            if ([expirationDate compare:now] == NSOrderedAscending) { // Expiration date has passed
+                [expiredObjectKeys addObject:key];
+            }
+        }];
+    [self unlock];
+
+    for (NSString *key in expiredObjectKeys) {
+        //unlock, removeFileAndExecuteBlocksForKey handles locking itself
+        [self removeFileAndExecuteBlocksForKey:key];
+    }
+}
+
 - (void)removeAllObjects
 {
     // We don't need to know the disk state since we're just going to remove everything.
@@ -1161,8 +1331,9 @@ static NSURL *_sharedTrashURL;
         for (NSString *key in _metadata) {
             NSURL *fileURL = [self encodedFileURLForKey:key];
             // If the cache should behave like a TTL cache, then only fetch the object if there's a valid ageLimit and the object is still alive
-            NSDate *date = _metadata[key].date;
-            if (!self->_ttlCache || self->_ageLimit <= 0 || (date && fabs([date timeIntervalSinceDate:now]) < self->_ageLimit)) {
+            NSDate *createdDate = _metadata[key].createdDate;
+            NSTimeInterval ageLimit = _metadata[key].ageLimit > 0.0 ? _metadata[key].ageLimit : self->_ageLimit;
+            if (!self->_ttlCache || ageLimit <= 0 || (createdDate && fabs([createdDate timeIntervalSinceDate:now]) < ageLimit)) {
                 BOOL stop = NO;
                 block(key, fileURL, &stop);
                 if (stop)