+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [fix] Fix up warnings and upgrade to PINOperation 1.1.1: [#213](https://github.com/pinterest/PINCache/pull/213)
 - [performance] Reduce locking churn in cleanup methods. [#212](https://github.com/pinterest/PINCache/pull/212)
 - [fix] Don't set file protection unless requested. [#220](https://github.com/pinterest/PINCache/pull/220)
+- [new] Add ability to set an object level TTL: [#209](https://github.com/pinterest/PINCache/pull/209)
 
 ## 3.0.1 -- Beta 6
 - [fix] Add some sane limits to the disk cache: [#201]https://github.com/pinterest/PINCache/pull/201